Install corrections

+++ nginx-admin-script/install/red-hat/install.sh
@@ -20,11 +20,16 @@ if ! rpm -q --quiet java ; then
   yum -y install java-1.8.0-openjdk-devel.x86_64
 fi
 mkdir -p /usr/share/softwares
-wget https://bintray.com/jslsolucoes/nginx-admin/download_file?file_path=nginx-admin-standalone-1.0.0-swarm.jar -O /usr/share/softwares/nginx-admin-standalone-1.0.0-swarm.jar
-wget https://raw.githubusercontent.com/jslsolucoes/nginx-admin/develop/nginx-admin-script/install/red-hat/nginx-admin-init-redhat.sh -O /etc/init.d/nginx-admin
-sed -i "s/dbuser \(\w*\)/dbpassword $dbuser/g" /etc/init.d/nginx-admin
-sed -i "s/dbpassword \(\w*\)/dbpassword $dbpassword/g" /etc/init.d/nginx-admin
-chmod +x /etc/init.d/nginx-admin
-chown root:root /etc/init.d/nginx-admin
-update-rc.d nginx-admin defaults
-update-rc.d nginx-admin enable
+
+if [! -f /usr/share/softwares/nginx-admin-standalone-1.0.0-swarm.jar] ; then 
+	wget https://bintray.com/jslsolucoes/nginx-admin/download_file?file_path=nginx-admin-standalone-1.0.0-swarm.jar -O /usr/share/softwares/nginx-admin-standalone-1.0.0-swarm.jar
+fi
+
+if [! -f /etc/init.d/nginx-admin] ; then 
+	wget https://raw.githubusercontent.com/jslsolucoes/nginx-admin/develop/nginx-admin-script/install/red-hat/nginx-admin-init-redhat.sh -O /etc/init.d/nginx-admin
+	sed -i "s/dbuser \(\w*\)/dbuser $dbuser/g" /etc/init.d/nginx-admin
+	sed -i "s/dbpassword \(\w*\)/dbpassword $dbpassword/g" /etc/init.d/nginx-admin
+	chmod +x /etc/init.d/nginx-admin
+	chown root:root /etc/init.d/nginx-admin
+	chkconfig --level 345 nginx-admin on
+fi
